% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{plot_confidence_intervals}
\alias{plot_confidence_intervals}
\title{Plotting function that samples at random a few estimated simulation runs
and shows an error plot highlighting
the size of the 95th confidence interval (black error bars) and whether the real value (red dots) are in
them.}
\usage{
plot_confidence_intervals(
  estimation,
  limit_number_of_runs_to_plot_to = 100,
  random_seed = 0
)
}
\arguments{
\item{estimation}{the output of one of the \code{fit_} or \code{cross_validate_} calls in this package}
}
\value{
A ggplot object
}
\description{
Useful to get a quick overview about the quality and size of the confidence intervals
}
\examples{
##generate some fake data where paramone,paramtwo ---> ssone,sswto;
## notice that paramtwo is basically unidentifiable!
paramone<-rnorm(n=5000)
paramtwo<-runif(n=5000,min=2,max=5)
ssone<-2*paramone + rnorm(n=5000)
sstwo<- paramone/paramtwo  + rnorm(n=5000)
training_data<-
  data.frame(
    paramone,
    paramtwo,
    ssone,
    sstwo
  )
  ## this would be the "real" data, what we want to estimate our model with!
testing_data<-data.frame(
  ssone=2,
 sstwo=0.25
)


### fit a gam
estimation<- fit_gam(training_runs = training_data,
                     target_runs =  training_data,
                     parameter_colnames = c("paramone","paramtwo"),
                     summary_statistics_colnames = c("ssone","sstwo"))
## we can check how the confidence intervals look like on a subset of runs
## because target_runs were just the training_data again this is showing IN-SAMPLE CI
plot_confidence_intervals(estimation)
## look in particular at how non-committal  paramtwo CIs are! Clearly GAM knows
## that it is NOT identifying that parameter

## we can do the same plot for cross-validations results; which
## would make it OUT-OF-SAMPLE
cv_results<-
 cross_validate_rejection_abc(training_data,ngroup = 5,
                              parameter_colnames = c("paramone","paramtwo"),
                              summary_statistics_colnames = c("ssone","sstwo"))
}
