% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{plot_grid_rmse}
\alias{plot_grid_rmse}
\title{Produces a tile plot showing for two chosen parameters the total root mean square error at different intervals
The idea here is to look at a 2D slice of a multi-parameter model to see if any particular space of the model
is better identified than the rest.}
\usage{
plot_grid_rmse(
  estimation,
  parameter1,
  parameter2,
  intervals_first_parameter = 10,
  intervals_second_parameter = 10
)
}
\arguments{
\item{estimation}{the output of one of the \code{fit_} or \code{cross_validate_} calls in this package}
}
\value{
A single ggplot
}
\description{
I find this plot particularly useful to spot when estimation is better/worse for
a sub-interval of the parameter range
#'
}
\examples{
##check the data scale
data("scale")
abc.cv<-cross_validate_rejection_abc(total_data = scale,
                                     parameter_colnames = c("weight1","weight2"),
                                     summary_statistics_colnames = c("total"))


}
