% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{plot_point_prediction_quality}
\alias{plot_point_prediction_quality}
\title{Produces a scatterplot of real parameters (y axis) and estimated parameters (x axis) to quickly
look at quality of point prediction. A perfect method would have all these dots on the 45 degree line
which is highlighted here with a red dashed line.}
\usage{
plot_point_prediction_quality(estimation)
}
\arguments{
\item{estimation}{the output of one of the \code{fit_} or \code{cross_validate_} calls in this package}
}
\value{
A list of ggplots
}
\description{
I find this plot particularly useful to spot partial identifications: when estimation is possible for
a sub-interval of the parameter range
#'
}
\examples{
##generate some fake data where paramone,paramtwo ---> ssone,sswto;
## notice that paramtwo is basically unidentifiable!
paramone<-rnorm(n=5000)
paramtwo<-runif(n=5000,min=2,max=5)
ssone<-2*paramone + rnorm(n=5000)
sstwo<- paramone/paramtwo  + rnorm(n=5000)
training_data<-
  data.frame(
    paramone,
    paramtwo,
    ssone,
    sstwo
  )
  ## this would be the "real" data, what we want to estimate our model with!
testing_data<-data.frame(
  ssone=2,
 sstwo=0.25
)


### fit a gam
estimation<- fit_gam(training_runs = training_data,
                     target_runs =  training_data,
                     parameter_colnames = c("paramone","paramtwo"),
                     summary_statistics_colnames = c("ssone","sstwo"))
## we can check how the prediction match real parameters
## because target_runs were just the training_data again this is showing IN-SAMPLE errors
plot_point_prediction_quality(estimation)
## notice that basically GAM for paramtwo just returns the average (i.e. "I don't know!")

## we can theme it, but we need to map (since this is patchwork output)
## we can do the same plot for cross-validations results; which
## would make it OUT-OF-SAMPLE
cv_results<-
 cross_validate_rejection_abc(training_data,ngroup = 5,
                              parameter_colnames = c("paramone","paramtwo"),
                              summary_statistics_colnames = c("ssone","sstwo"))
}
