% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_experiment-data.R
\docType{data}
\name{scale_experiment}
\alias{scale_experiment}
\title{Output from 5,000 replications of the scale experiment: two people stand on a scale and we only record their total
We vary two parameters \code{weight1}, \code{weight2} and collect a single summary statistic \code{total}}
\format{
An object of class \code{"tibble"};
}
\usage{
data(scale_experiment)
}
\description{
Output from 5,000 replications of the scale experiment: two people stand on a scale and we only record their total
We vary two parameters \code{weight1}, \code{weight2} and collect a single summary statistic \code{total}
}
\references{
Ernesto Carrella (under revision). No free lunch when estimating simulation parameters
}
\keyword{datasets}
