% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{tidy_up_estimation}
\alias{tidy_up_estimation}
\title{Helper function to produce a tidy data frame containing upper,lower,estimate and real value for each parameter
in each simulated run. Useful for plotting}
\usage{
tidy_up_estimation(estimation)
}
\arguments{
\item{estimation}{the output of one of the \code{fit_} or \code{cross_validate_} calls in this package}
}
\value{
A data-frame containing the following columns
 \itemize{
  \item type whether the row is the estimated parameter, its bounds or the real value
  \item run the simulation run index
  \item variable which parameter are we estimating
  \item value the estimated value (NA if we don't know the REAL value of a parameter when type is "real")
}
}
\description{
Helper function to produce a tidy data frame containing upper,lower,estimate and real value for each parameter
in each simulated run. Useful for plotting
}
\examples{
##generate some fake data where paramone,paramtwo ---> ssone,sswto;
## notice that paramtwo is basically unidentifiable!
paramone<-rnorm(n=5000)
paramtwo<-runif(n=5000,min=2,max=5)
ssone<-2*paramone + rnorm(n=5000)
sstwo<- paramone/paramtwo  + rnorm(n=5000)
training_data<-
  data.frame(
    paramone,
    paramtwo,
    ssone,
    sstwo
  )
  ## this would be the "real" data, what we want to estimate our model with!
testing_data<-data.frame(
  ssone=2,
 sstwo=0.25
)


### fit a gam
estimation<- fit_gam(training_runs = training_data,
                     target_runs =  training_data,
                     parameter_colnames = c("paramone","paramtwo"),
                     summary_statistics_colnames = c("ssone","sstwo"))

## we can clean it to be more presentable
tidy_up_estimation(estimation)

## we can do the same with a cv result
cv_results<-
cross_validate_rejection_abc(training_data,ngroup = 5,
                            parameter_colnames = c("paramone","paramtwo"),
                            summary_statistics_colnames = c("ssone","sstwo"))
tidy_up_estimation(cv_results)
}
